23032022
changed R from int to float

+++ circulo.java
@@ -2,8 +2,8 @@ package Unidad3;
 
 public class circulo extends Elipse {
 
-	private int R;
-	public circulo(int R, int D, String nombre) {
+	private float R;
+	public circulo(float R, int D, String nombre) {
 		this.R = R;
 	}
